# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Exclusive Job Portal API Documentation",
        "version": "0.1"
    },
    "paths": {
        "/api/exams/": {
            "get": {
                "tags": [
                    "Exam"
                ],
                "summary": "Get exams details",
                "description": "Returns exam information",
                "operationId": "a5774d58a3f849c00d5fe3b9284ef659",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/exams/questions/{examId}": {
            "get": {
                "tags": [
                    "Exam"
                ],
                "summary": "Get exams details",
                "description": "Returns exam's questions information",
                "operationId": "44cb84845f245f66ae03fadea3930373",
                "parameters": [
                    {
                        "name": "examId",
                        "in": "path",
                        "description": "Exam unique ID",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/exams/questions-answers/score": {
            "post": {
                "tags": [
                    "Exam"
                ],
                "summary": "Store exam score",
                "description": "Returns score number",
                "operationId": "7a3385a262543bde465c87f07857ae2e",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "parentId",
                                    "questions[]",
                                    "answers[]"
                                ],
                                "properties": {
                                    "parentId": {
                                        "type": "integer"
                                    },
                                    "questions[]": {
                                        "description": "Questions ID",
                                        "type": "array",
                                        "items": {
                                            "type": "int",
                                            "format": "id"
                                        },
                                        "collectionFormat": "multi"
                                    },
                                    "answers[]": {
                                        "description": "Answers ID",
                                        "type": "array",
                                        "items": {
                                            "type": "int",
                                            "format": "id"
                                        },
                                        "collectionFormat": "multi"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login Successful"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/packages/": {
            "get": {
                "tags": [
                    "Package"
                ],
                "summary": "Get package details",
                "description": "Returns package information",
                "operationId": "bc7fc82a6159321d499f8a13a4ed144b",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/packages/courses/{packageId}": {
            "get": {
                "tags": [
                    "Package"
                ],
                "summary": "Get packages course details",
                "description": "Returns packages's courses information",
                "operationId": "c35910c6bedee5d623f2f77be290a892",
                "parameters": [
                    {
                        "name": "packageId",
                        "in": "path",
                        "description": "Package unique ID",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/packages/sections/{packageId}": {
            "get": {
                "tags": [
                    "Package"
                ],
                "summary": "Get packages section details",
                "description": "Returns packages's sections information",
                "operationId": "335bae975c0755598ce4f0b9ef44f283",
                "parameters": [
                    {
                        "name": "packageId",
                        "in": "path",
                        "description": "Package unique ID",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    }
}